UPDATE: resources calculator add melee and secondary
