--- src/components/masteryCalculator/Recommends.vue
@@ -1,13 +1,0 @@
-<template>
-    <b-card header="Recommendations">
-        <p>123</p>
-    </b-card>
-</template>
-
-<script>
-export default {
-    name: 'recommends'
-};
-</script>
-
-<style scoped></style>
